# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-commerce REST API with products, categories, orders, users, and inventory management",
        "title": "Bey API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@bey.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the authenticated user's shopping cart items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Get shopping cart",
                "responses": {
                    "200": {
                        "description": "Cart retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_cart.CartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes all items from the authenticated user's shopping cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Clear shopping cart",
                "responses": {
                    "200": {
                        "description": "Cart cleared successfully",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/cart/items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a product variant to the authenticated user's shopping cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Add item to cart",
                "parameters": [
                    {
                        "description": "Item to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_cart.AddToCartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item added to cart",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_cart.CartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid data or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Variant not found",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/cart/items/{variant_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the quantity of an item in the authenticated user's shopping cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Update cart item quantity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product variant ID",
                        "name": "variant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New quantity",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_cart.UpdateCartItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_cart.CartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid data or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Item not found in cart",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a specific item from the authenticated user's shopping cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Remove item from cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product variant ID to remove",
                        "name": "variant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_cart.CartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid variant ID",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bey_internal_shared_response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_modules_cart.AddToCartRequest": {
            "type": "object",
            "required": [
                "quantity",
                "variant_id"
            ],
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "variant_id": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_cart.CartItemResponse": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "variant_id": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_cart.CartResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_cart.CartItemResponse"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_cart.UpdateCartItemRequest": {
            "type": "object",
            "required": [
                "quantity"
            ],
            "properties": {
                "quantity": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your Bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}